+++ src/hooks/use-practiceExam-context.tsx
@@ -1,0 +1,8 @@
+import {useContext} from 'react'
+import PracticeExamContext from '../context/PracticeExamContext'
+
+function usePracticeExamContext() {
+    return useContext(PracticeExamContext);
+}
+
+export default usePracticeExamContext;
