+++ src/models/QuestionType.tsx
@@ -1,0 +1,6 @@
+export type Question = {
+  question: string;
+  answer: string;
+  options: string[];
+  explanation: string;
+}
